Rename typechecking file
